+++ src/components/Buy/index.js
@@ -3,12 +3,16 @@ import { Section, Wrapper, Heading } from 'common/base';
 import SubscriptionsSection from './SubscriptionsSection';
 import PaymentSection from './PaymentSection';
 
-export default () => (
-  <Section paddingTop paddingBottom>
-    <Wrapper size="s">
-      <Heading center>付費解鎖全站</Heading>
-      <SubscriptionsSection />
-      <PaymentSection />
-    </Wrapper>
-  </Section>
-);
+const Buy = () => {
+  return (
+    <Section paddingTop paddingBottom>
+      <Wrapper size="s">
+        <Heading center>付費解鎖全站</Heading>
+        <SubscriptionsSection />
+        <PaymentSection />
+      </Wrapper>
+    </Section>
+  );
+};
+
+export default Buy;
